fix(resolver): correct compose provenance attribution
