+++ packages/browser-compiler/src/resolver.ts
@@ -30,6 +30,7 @@ import {
   normalizeProgram,
   collectProvenance,
   collectProvenanceEvents,
+  collectCompositionProvenanceEvents,
   collectProvenanceValueEvents,
   emptyProvenance,
   prefixProvenance,
@@ -235,6 +236,10 @@ function effectiveCompositionPath(blocks: readonly Block[], path: string): strin
     return parts.slice(1).join('.');
   }
   return path;
+}
+
+function isValueRecord(value: unknown): value is Record<string, Value> {
+  return typeof value === 'object' && value !== null && !Array.isArray(value);
 }
 
 // ── Skill-aware merge strategy sets ──────────────────────────────────
@@ -492,21 +497,19 @@ export class BrowserResolver {
       this.logger.debug(`Applying ${ast.extends.length} extension(s)`);
     }
     if (!sequentialOperations) {
-      const extensionBlocks = ast.blocks;
-      for (const extension of ast.extends) {
-        provenanceEvents.push(
-          ...collectProvenanceEvents(
-            extension.canonicalBody,
-            effectiveCompositionPath(extensionBlocks, extension.targetPath),
-            'extend',
-            extension.loc,
-            extension.replacements?.length ? 'replaced' : 'merged',
-            extension.replacements?.length ? 'replace' : 'merge'
-          )
-        );
-      }
       try {
-        ast = this.applyExtends(ast);
+        ast = this.applyExtends(ast, (extension, extensionBlocks) => {
+          provenanceEvents.push(
+            ...collectProvenanceEvents(
+              extension.canonicalBody,
+              effectiveCompositionPath(extensionBlocks, extension.targetPath),
+              'extend',
+              extension.loc,
+              extension.replacements?.length ? 'replaced' : 'merged',
+              extension.replacements?.length ? 'replace' : 'merge'
+            )
+          );
+        });
       } catch (err) {
         if (err instanceof ResolveError) {
           errors.push(err);
@@ -1012,66 +1015,17 @@ export class BrowserResolver {
       );
       if (skillsBlock?.content.type === 'ObjectContent') {
         for (const [skillName, skillValue] of Object.entries(skillsBlock.content.properties)) {
-          if (typeof skillValue !== 'object' || skillValue === null || Array.isArray(skillValue)) {
-            continue;
-          }
-          const composed = (skillValue as Record<string, Value>)['__composedFrom'];
+          if (!isValueRecord(skillValue)) continue;
+          const composed = skillValue['__composedFrom'];
           if (!Array.isArray(composed)) continue;
-          const usedInlineUses = new Set<number>();
-          for (const [index, phase] of composed.entries()) {
-            if (
-              typeof phase !== 'object' ||
-              phase === null ||
-              Array.isArray(phase) ||
-              typeof (phase as Record<string, unknown>)['source'] !== 'string'
-            ) {
-              continue;
-            }
-            const sourceFile = (phase as Record<string, unknown>)['source'] as string;
-            const matchedIndex = inlineUses.findIndex(({ declaration }, candidateIndex) => {
-              if (usedInlineUses.has(candidateIndex)) return false;
-              try {
-                return this.resolveRef(declaration.path, absPath) === sourceFile;
-              } catch {
-                return false;
-              }
-            });
-            const useIndex = matchedIndex >= 0 ? matchedIndex : index;
-            usedInlineUses.add(useIndex);
-            const use = inlineUses[useIndex]?.declaration;
-            if (!use) continue;
-            const source = {
-              file: sourceFile,
-              line: 1,
-              column: 1,
-              offset: 0,
-            };
-            const chain: ProvenanceLink = {
-              operation: 'compose',
-              source: deepClone(use.loc),
-              target: source.file,
-              reference: use.path.raw,
-              ...(use.alias ? { alias: use.alias } : {}),
-            };
-            for (const property of [
-              'content',
-              'allowedTools',
-              'references',
-              'requires',
-              'inputs',
-              'outputs',
-            ]) {
-              provenanceEvents.push({
-                path: `skills.${skillName}.${property}`,
-                operation: 'compose',
-                action: 'composed',
-                source,
-                target: source.file,
-                reference: use.path.raw,
-                chain: [chain],
-              });
-            }
-          }
+          provenanceEvents.push(
+            ...collectCompositionProvenanceEvents(
+              composed,
+              inlineUses,
+              (declaration) => this.resolveRef(declaration.path, absPath),
+              skillName
+            )
+          );
         }
       }
     } catch (err) {
@@ -1136,11 +1090,16 @@ export class BrowserResolver {
   /**
    * Apply all @extend blocks to resolve extensions.
    */
-  private applyExtends(ast: Program): Program {
+  private applyExtends(
+    ast: Program,
+    onExtensionApplied?: (extension: ExtendBlock, previousBlocks: readonly Block[]) => void
+  ): Program {
     let blocks = [...ast.blocks];
 
     for (const ext of ast.extends) {
+      const previousBlocks = blocks;
       blocks = this.applyExtend(blocks, ext);
+      onExtensionApplied?.(ext, previousBlocks);
     }
 
     blocks = blocks.filter((b) => !b.name.startsWith(IMPORT_MARKER_PREFIX));
