+++ packages/resolver/src/skill-composition.ts
@@ -8,6 +8,7 @@ import type {
   InlineUseDeclaration,
   ComposedPhase,
   SkillContractField,
+  SourceLocation,
   SyntaxFeatureUsage,
 } from '@promptscript/core';
 import { deepClone, getSyntaxFeatureUsages, isTextContent, ResolveError } from '@promptscript/core';
@@ -189,6 +190,8 @@ interface ResolvedPhase {
   name: string;
   /** Absolute source path */
   source: string;
+  /** Source location of the composed skill definition */
+  loc: SourceLocation;
   /** Alias if specified */
   alias?: string;
   /** Skill description */
@@ -287,6 +290,7 @@ async function resolveInlineUse(
     phaseNumber,
     name: phaseName,
     source: absPath,
+    loc: inlineUse.loc,
     alias: inlineUse.alias,
     description: skillDef.description,
     skillContent: skillDef.content,
@@ -488,6 +492,7 @@ function composeIntoSkill(
     const composedPhase: ComposedPhase = {
       name: phase.name,
       source: phase.source,
+      loc: phase.loc,
       composedBlocks,
     };
